LoRa/GUI/GUIv5/requirements.txt
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyQt5-stubs==5.15.2.0
pyserial==3.5
